requirements.txt
flask
flask-cors
gunicorn
torch
transformers
pillow
requests